refactor: separate Cdf and local point cloud factories
