+++ viewer/packages/pointclouds/src/PointCloudMetadata.ts
@@ -4,13 +4,14 @@
 
 import * as THREE from 'three';
 import { CameraConfiguration } from '@reveal/utilities';
-import { File3dFormat } from '@reveal/modeldata-api';
+import { File3dFormat, ModelIdentifier } from '@reveal/modeldata-api';
 
 export interface PointCloudMetadata {
   readonly format: File3dFormat;
   readonly formatVersion: number;
 
   readonly modelBaseUrl: string;
+  readonly modelIdentifier: ModelIdentifier;
   readonly modelMatrix: THREE.Matrix4;
   readonly cameraConfiguration?: CameraConfiguration;
   readonly scene: any;
