+++ viewer/packages/pointclouds/src/createPointCloudManager.ts
@@ -4,13 +4,14 @@
 
 import { PointCloudManager } from './PointCloudManager';
 import { PointCloudMetadataRepository } from './PointCloudMetadataRepository';
-import { PointCloudFactory } from './PointCloudFactory';
+import { CdfPointCloudFactory } from './factory/CdfPointCloudFactory';
 
 import * as THREE from 'three';
 
 import { ModelDataProvider, ModelMetadataProvider } from '@reveal/modeldata-api';
 import { CogniteClient } from '@cognite/sdk';
 import { Potree } from './potree-three-loader';
+import { LocalPointCloudFactory } from './factory/LocalPointCloudFactory';
 
 export function createPointCloudManager(
   modelMetadataProvider: ModelMetadataProvider,
@@ -21,6 +22,12 @@ export function createPointCloudManager(
 ): PointCloudManager {
   const metadataRepository = new PointCloudMetadataRepository(modelMetadataProvider, modelDataProvider);
   const potreeInstance = new Potree(modelDataProvider);
-  const modelFactory = new PointCloudFactory(potreeInstance, sdkClient);
-  return new PointCloudManager(metadataRepository, modelFactory, scene, renderer);
+
+  let modelFactory;
+  if (sdkClient) {
+    modelFactory = new CdfPointCloudFactory(potreeInstance, sdkClient);
+  } else {
+    modelFactory = new LocalPointCloudFactory(potreeInstance);
+  }
+  return new PointCloudManager(metadataRepository, modelFactory, potreeInstance, scene, renderer);
 }
